getting url piece as a parameter
