+++ src/scenes/Office/QueueTable.jsx
@@ -1,48 +1,8 @@
 import React, { Component } from 'react';
 import { render } from 'react-dom';
-import _ from 'lodash';
-import { FunctionNameGoesHere } from './api.js';
-
 import ReactTable from 'react-table';
 import 'react-table/react-table.css';
-
-const rawData = FunctionNameGoesHere();
-
-const requestData = (pageSize, page, sorted, filtered) => {
-  return new Promise((resolve, reject) => {
-    let filteredData = rawData;
-
-    // You can use the filters in your request, but you are responsible for applying them.
-    if (filtered.length) {
-      filteredData = filtered.reduce((filteredSoFar, nextFilter) => {
-        return filteredSoFar.filter(row => {
-          return (row[nextFilter.id] + '').includes(nextFilter.value);
-        });
-      }, filteredData);
-    }
-    // You can also use the sorting in your request, but again, you are responsible for applying it.
-    const sortedData = _.orderBy(
-      filteredData,
-      sorted.map(sort => {
-        return row => {
-          if (row[sort.id] === null || row[sort.id] === undefined) {
-            return -Infinity;
-          }
-          return typeof row[sort.id] === 'string'
-            ? row[sort.id].toLowerCase()
-            : row[sort.id];
-        };
-      }),
-      sorted.map(d => (d.desc ? 'desc' : 'asc')),
-    );
-
-    // You must return an object containing the rows of the current page, and optionally the total pages number.
-    const res = {
-      rows: filteredData.slice(pageSize * page, pageSize * page + pageSize),
-      pages: Math.ceil(filteredData.length / pageSize),
-    };
-  });
-};
+import { requestData } from './ducks.js';
 
 export default class QueueTable extends Component {
   constructor() {
@@ -60,16 +20,20 @@ export default class QueueTable extends Component {
     // You can set the `loading` prop of the table to true to use the built-in one or show you're own loading bar if you want.
     this.setState({ loading: true });
     // Request the data however you want.  Here, we'll use our mocked service we created earlier
-    requestData(state.pageSize, state.page, state.sorted, state.filtered).then(
-      res => {
-        // Now just get the rows of data to your React Table (and update anything else like total pages or loading)
-        this.setState({
-          data: res.rows,
-          pages: res.pages,
-          loading: false,
-        });
-      },
-    );
+    requestData(
+      this.props.queueType,
+      state.pageSize,
+      state.page,
+      state.sorted,
+      state.filtered,
+    ).then(res => {
+      // Now just get the rows of data to your React Table (and update anything else like total pages or loading)
+      this.setState({
+        data: res.rows,
+        pages: res.pages,
+        loading: false,
+      });
+    });
   }
 
   render() {
