+++ src/scenes/Office/api.js
@@ -5,11 +5,11 @@ export async function GetSpec() {
   return client.spec;
 }
 
-export async function FunctionNameGoesHere(payload) {
-  // Pull from finished API definition
+export async function RetrieveMovesForOffice(queueType) {
   const client = await getClient();
-  const response = await client.apis.duty_stations.searchDutyStations({
-    status: 0, // Parameters to come; how many to add?
+  debugger;
+  const response = await client.apis.queues.showQueue({
+    queueType,
   });
   checkResponse(response, 'failed to retrieve moves due to server error');
   return response.body;
